Moved PriceUpdateBody struct to Content/PriceUpdateBody.swift file
