+++ Sources/App/Models/Price.swift
@@ -82,23 +82,6 @@ final class Price: Content, MySQLModel, Parameter {
         // This method is also used by Fluent to create the tables in the datbase.
         self.id = try container.decodeIfPresent(Int.self, forKey: .id)
     }
-    
-    
-    /// Updates the model with data from a request and saves it.
-    ///
-    /// - Parameters:
-    ///   - body: The body of a request, decoded to a `PriceUpdateBody`.
-    ///   - executor: The object that will be used to save the model to the database.
-    /// - Returns: A void future, which will signal once the update is complete.
-    func update(with body: PriceUpdateBody) -> Price {
-        // Update all the properties if a value for it is found in the body, else use the old value.
-        self.cents = body.cents ?? self.cents
-        self.activeFrom = body.activeFrom ?? self.activeFrom
-        self.activeTo = body.activeTo ?? self.activeTo
-        self.active = body.active ?? self.active
-        
-        return self
-    }
 }
 
 extension Price: Migration {
@@ -113,20 +96,4 @@ extension Price: Migration {
             builder.reference(from: \.productID, to: \Product.id)
         }
     }
-}
-
-/// A representation of a request's body when you need to update a `Price` model.
-struct PriceUpdateBody: Content {
-    
-    ///
-    let cents: Int?
-    
-    ///
-    let activeFrom: Date?
-    
-    ///
-    let activeTo: Date?
-    
-    ///
-    let active: Bool?
 }
